+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3890,7 +3890,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M3BY48XECV4HVVKETAQTFKMK"),
+  id: z.string().default("01M3CFSBTNR3SKD8HMSTWE1MYS"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),