+++ packages/orq-rc/RELEASES.md
@@ -12639,3 +12639,13 @@ Based on:
 - [typescript v4.16.0-rc.19] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.19 - packages/orq-rc
+
+## 2026-09-25 14:37:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.799.0 (2.941.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.20] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.20 - packages/orq-rc